+++ HBase/src/main/java/com/hzgc/hbase/dynamicrepo/CaptureHistory.java
@@ -3,8 +3,8 @@ package com.hzgc.hbase.dynamicrepo;
 import com.hzgc.dubbo.attribute.Attribute;
 import com.hzgc.dubbo.attribute.AttributeValue;
 import com.hzgc.dubbo.dynamicrepo.*;
+import com.hzgc.ftpserver.util.FtpUtil;
 import com.hzgc.hbase.staticrepo.ElasticSearchHelper;
-import com.hzgc.hbase.util.FtpUtil;
 import org.apache.log4j.Logger;
 import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.action.search.SearchResponse;
