+++ HBase/src/main/java/com/hzgc/hbase/dynamicrepo/CapturePictureSearchServiceImpl.java
@@ -325,6 +325,7 @@ public class CapturePictureSearchServiceImpl implements CapturePictureSearchServ
     @Override
     public SearchResult getCaptureHistory(SearchOption option) {
         CaptureHistory captureHistory = new CaptureHistory();
+        option.setSearchType(SearchType.PERSON);
         long esStartTime = System.currentTimeMillis();
         SearchResult searchResult = captureHistory.getRowKey_history(option);
         long esEndTime = System.currentTimeMillis();
